+++ NEWS.md
@@ -1,5 +1,7 @@
 # EDBdata (development version)
 
+* Added `phyto_edb` - phytoplankton community data collected by DWR-EMP from 2014-2021 for the stations within the designated EDB regions
+
 # EDBdata 0.1.0
 
 * Finished processing HAB satellite data for Franks Tract and Mildred Island which is in the `sat_ci_count_fr_mil` data object. This is the data used in the November draft of the EDB report.